Move timeline / profiler colors to separate file for sharing.
