+++ packages/devtools/lib/src/timeline/timeline_screen.dart
@@ -11,7 +11,6 @@ import '../globals.dart';
 import '../ui/analytics.dart' as ga;
 import '../ui/analytics_platform.dart' as ga_platform;
 import '../ui/elements.dart';
-import '../ui/fake_flutter/dart_ui/dart_ui.dart';
 import '../ui/icons.dart';
 import '../ui/material_icons.dart';
 import '../ui/primer.dart';
@@ -23,39 +22,6 @@ import 'frames_bar_chart.dart';
 import 'timeline_controller.dart';
 import 'timeline_model.dart';
 import 'timeline_protocol.dart';
-
-// Light mode is Light Blue 50 palette and Dark mode is Blue 50 palette.
-// https://material.io/design/color/the-color-system.html#tools-for-picking-colors.
-const mainUiColorLight = Color(0xFF81D4FA); // Light Blue 50 - 200
-const mainUiColorSelectedLight = Color(0xFFD4D7DA); // Lighter grey.
-
-const mainGpuColorLight = Color(0xFF0288D1); // Light Blue 50 - 700
-const mainGpuColorSelectedLight = Color(0xFFB5B5B5); // Darker grey.
-
-const mainUiColorDark = Color(0xFF9EBEF9); // Blue 200 Material Dark
-const mainUiColorSelectedDark = Colors.white;
-
-const mainGpuColorDark = Color(0xFF1A73E8); // Blue 600 Material Dark
-const mainGpuColorSelectedDark = Color(0xFFC9C9C9); // Grey.
-
-const mainUiColor = ThemedColor(mainUiColorLight, mainUiColorDark);
-const mainGpuColor = ThemedColor(mainGpuColorLight, mainGpuColorDark);
-
-const Color selectedUiColor =
-    ThemedColor(mainUiColorSelectedLight, mainUiColorSelectedDark);
-const Color selectedGpuColor =
-    ThemedColor(mainGpuColorSelectedLight, mainGpuColorSelectedDark);
-
-// Light is Red @ .2 opacity, Dark is Red 200 Material Dark @ .2 opacity.
-const Color jankGlowInside = ThemedColor(Color(0x66FF0000), Color(0x66F29C99));
-// Light is Red @ .5 opacity, Dark is Red 600 Material Dark @ .6 opacity.
-const Color jankGlowEdge = ThemedColor(Color(0x80FF0000), Color(0x99CE191C));
-
-// Red 50 - 400 is light at 1/2 opacity, Dark Red 500 Material Dark.
-const Color highwater16msColor = mainUiColorSelectedLight;
-
-const Color hoverTextHighContrastColor = Colors.white;
-const Color hoverTextColor = Colors.black;
 
 // TODO(devoncarew): show the Skia picture (gpu drawing commands) for a frame
